backend/requirements.txt
asgiref==3.7.2
Django==4.2.9
Pillow==10.2.0
# pymysql==1.1.0
# mysqlclient
gunicorn==21.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
requests==2.31.0
django-cors-headers==4.3.1
daphne==4.1.2
django-mptt==0.16.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.1
django-debug-toolbar==4.3.0
django-tinymce==4.0.0
flake8==7.0.0
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.3
channels-redis==4.2.0


